migrate mailbox to lit
